Move one of the tests to integration for full code coverage
