+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip551/AtomicBatchIntegrationTest.java
@@ -1,94 +1,100 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.hedera.services.bdd.suites.hip551;
 
+import static com.google.protobuf.ByteString.copyFromUtf8;
 import static com.hedera.services.bdd.junit.TestTags.INTEGRATION;
 import static com.hedera.services.bdd.junit.hedera.embedded.EmbeddedMode.CONCURRENT;
 import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
 import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.resultWith;
 import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.recordWith;
 import static com.hedera.services.bdd.spec.keys.KeyShape.CONTRACT;
+import static com.hedera.services.bdd.spec.keys.KeyShape.PREDEFINED_SHAPE;
 import static com.hedera.services.bdd.spec.keys.KeyShape.sigs;
-import static com.hedera.services.bdd.spec.keys.SigControl.ON;
 import static com.hedera.services.bdd.spec.queries.QueryVerbs.getTxnRecord;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.atomicBatch;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCall;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.contractCreate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.cryptoCreate;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.mintToken;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenCreate;
+import static com.hedera.services.bdd.spec.transactions.TxnVerbs.tokenUpdate;
 import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
 import static com.hedera.services.bdd.spec.utilops.CustomSpecAssert.allRunFor;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.accountAmount;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.childRecordsCheck;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sourcing;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.transferList;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
-import static com.hedera.services.bdd.suites.HapiSuite.ONE_HBAR;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HUNDRED_HBARS;
+import static com.hedera.services.bdd.suites.contract.precompile.ContractBurnHTSSuite.ALICE;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
 
 import com.esaulpaugh.headlong.abi.Address;
-import com.esaulpaugh.headlong.abi.Tuple;
 import com.hedera.services.bdd.junit.HapiTest;
+import com.hedera.services.bdd.junit.HapiTestLifecycle;
 import com.hedera.services.bdd.junit.TargetEmbeddedMode;
+import com.hedera.services.bdd.junit.support.TestLifecycle;
 import com.hedera.services.bdd.spec.keys.KeyShape;
+import com.hederahashgraph.api.proto.java.TokenType;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Supplier;
 import java.util.stream.Stream;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.Tag;
 
 @Tag(INTEGRATION)
 @TargetEmbeddedMode(CONCURRENT)
+@HapiTestLifecycle
 public class AtomicBatchIntegrationTest {
 
+    @BeforeAll
+    static void beforeAll(@NonNull final TestLifecycle testLifecycle) {
+        testLifecycle.overrideInClass(
+                Map.of("atomicBatch.isEnabled", "true", "atomicBatch.maxNumberOfTransactions", "50"));
+    }
+
     @HapiTest
-    @DisplayName("Validate crypto transfer precompile gas used for inner transaction")
-    final Stream<DynamicTest> validateInnerCallToCryptoTransferPrecompile() {
-        final var sender = "sender";
-        final var receiver = "receiver";
+    @DisplayName("Validate burn precompile gas used for inner transaction")
+    final Stream<DynamicTest> validateInnerCallToBurnPrecompile() {
+        final var nft = "nft";
         final var gasToOffer = 2_000_000L;
-        final var transferContract = "AtomicCryptoTransfer";
+        final var burnContract = "BurnToken";
+        final var supplyKey = "supplyKey";
+        final AtomicReference<Address> tokenAddress = new AtomicReference<>();
+        final KeyShape listOfPredefinedAndContract = KeyShape.threshOf(1, PREDEFINED_SHAPE, CONTRACT);
         final AtomicLong gasUsed = new AtomicLong(0);
-        final KeyShape simpleContractKeyShape = KeyShape.threshOf(1, KeyShape.SIMPLE, CONTRACT);
-
-        final AtomicReference<Address> senderAddress = new AtomicReference<>();
-        final AtomicReference<Address> receiverAddress = new AtomicReference<>();
-
-        // call parameters
-        final Supplier<Tuple> transferListSupplier = () -> transferList()
-                .withAccountAmounts(
-                        accountAmount(senderAddress.get(), -ONE_HBAR, false),
-                        accountAmount(receiverAddress.get(), ONE_HBAR, false))
-                .build();
-        final var EMPTY_TUPLE_ARRAY = new Tuple[] {};
-
         return hapiTest(
-                // deploy the contract
-                uploadInitCode(transferContract),
-                contractCreate(transferContract).gas(gasToOffer),
-
-                // create sender and receiver with proper keys and expose their addresses
-                newKeyNamed("key").shape(simpleContractKeyShape.signedWith(sigs(ON, transferContract))),
-                cryptoCreate(sender).key("key").balance(ONE_HUNDRED_HBARS).exposingEvmAddressTo(senderAddress::set),
-                cryptoCreate(receiver).key("key").balance(0L).exposingEvmAddressTo(receiverAddress::set),
-                cryptoCreate("operator"),
-
-                // Simple transfer between sender, receiver
-                sourcing(() -> contractCall(
-                                transferContract,
-                                "transferMultipleTokens",
-                                transferListSupplier.get(),
-                                EMPTY_TUPLE_ARRAY)
-                        .via("cryptoTransferTxn")
+                cryptoCreate(ALICE).balance(10 * ONE_HUNDRED_HBARS),
+                tokenCreate(nft)
+                        .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
+                        .initialSupply(0L)
+                        .supplyKey(ALICE)
+                        .adminKey(ALICE)
+                        .treasury(ALICE)
+                        .exposingAddressTo(tokenAddress::set),
+                mintToken(nft, List.of(copyFromUtf8("1"), copyFromUtf8("2"))),
+                uploadInitCode(burnContract),
+                sourcing(() -> contractCreate(burnContract, tokenAddress.get())
+                        .payingWith(ALICE)
                         .gas(gasToOffer)),
+                newKeyNamed(supplyKey).shape(listOfPredefinedAndContract.signedWith(sigs(ALICE, burnContract))),
+                tokenUpdate(nft).supplyKey(supplyKey).signedByPayerAnd(ALICE),
+
+                // burn NFT via precompile and save the used gas
+                contractCall(burnContract, "burnToken", BigInteger.ZERO, new long[] {1L})
+                        .payingWith(ALICE)
+                        .alsoSigningWithFullPrefix(supplyKey)
+                        .gas(gasToOffer)
+                        .via("burn"),
 
                 // save precompile gas used
                 withOpContext((spec, op) -> {
-                    final var callRecord = getTxnRecord("cryptoTransferTxn")
-                            .andAllChildRecords()
-                            .logged();
+                    final var callRecord = getTxnRecord("burn").andAllChildRecords();
                     allRunFor(spec, callRecord);
                     gasUsed.set(callRecord
                             .getFirstNonStakingChildRecord()
@@ -96,20 +102,18 @@ public class AtomicBatchIntegrationTest {
                             .getGasUsed());
                 }),
 
-                // transfer hbars via precompile as inner batch txn
-                sourcing(() -> atomicBatch(contractCall(
-                                        transferContract,
-                                        "transferMultipleTokens",
-                                        transferListSupplier.get(),
-                                        EMPTY_TUPLE_ARRAY)
-                                .batchKey("operator")
-                                .via("cryptoTransferFromBatch")
-                                .gas(gasToOffer))
-                        .payingWith("operator")),
+                // burn NFT via precompile as inner batch txn
+                atomicBatch(contractCall(burnContract, "burnToken", BigInteger.ZERO, new long[] {2L})
+                                .batchKey(ALICE)
+                                .payingWith(ALICE)
+                                .alsoSigningWithFullPrefix(supplyKey)
+                                .gas(gasToOffer)
+                                .via("burnFromBatch"))
+                        .payingWith(ALICE),
 
                 // validate precompile used gas is the same as in the previous call
                 sourcing(() -> childRecordsCheck(
-                        "cryptoTransferFromBatch",
+                        "burnFromBatch",
                         SUCCESS,
                         recordWith()
                                 .status(SUCCESS)
